refactor(win32): extract dispatcher
